Spec replacing constraints
